Put the instances on screen
Not styled, and I'm cheating with the example text.
But it's a start.

I've swapped to preparing the template values, then applying them to the
template. `horrorshow` didn't work well with returning errors from
inside it, or with rendering sections ahead and adding it into the
document. Plus this feels more adaptable to other libraries if I decide
to switch like https://crates.io/crates/askama