Concentrate on paint values
I want to be able to get proper background values (not just colours).

I suspect this documentation will come in handy
https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Images/Using_CSS_gradients
https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Backgrounds_and_Borders/Using_multiple_backgrounds

I will also try converting paint structs to typescript with typeshare
https://crates.io/crates/typeshare

Last time I tried this there were a couple problems:
1. The flattened enums (which I'm getting rid of)
2. The fixed size arrays, which I'm keeping

I'll try and submit a fix for the 2nd

This may be a Rust library, but I'm seeing a lot of the value is in the
types- they should be shared to other languages.