Implement outlines
They're a bit of a pain.
I initially thought to do them as borders, as that's what I'm more
familiar with. But the border adds to the width and height.
`visible width = padding + border`. This isn't affected by `box-sizing`
as we're not setting the width or height directly.

Outline isn't used in the visible width or height, so it is the better
option to use.