Try again at uploading artifacts
Previously it didn't work as they were a directory.

I'm guessing at the actual file paths.

https://github.com/ccouzens/figma-rust/actions/runs/4908896654

```
Artifact is a directory:figma-file-tools-X64-Linux/. Directories can not be uploaded to a release.
Artifact is a directory:figma-file-tools-X64-macOS/. Directories can not be uploaded to a release.
Artifact is a directory:figma-file-tools-X64-Windows.exe/. Directories can not be uploaded to a release.
```